refactor: 코드 리팩토링 (#77)
- barrel 정리
