+++ src/app/(with-layout)/my-page/edit/page.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { Button, AlarmModal, EditNickNameModal } from '@/components'
-import ResignationModal from '@/components/common/modal/ResignationModal'
+import { ResignationModal } from '@/components/common'
 import { useEditPage, useEditPageModals } from '@/hooks/my-page'
 import { useLoginStore } from '@/stores/useLoginStore'
 import { getUsername } from '@/utils/getUsername'
