+++ src/app/(with-layout)/my-page/page.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { Button, RecommendedModal } from '@/components'
-import SelectCartModal from '@/components/common/modal/SelectCartModal'
+import { SelectCartModal } from '@/components/common'
 import { TasteTypeResult, TasteFingerprint } from '@/components/my-page'
 import { useRecommendedModal } from '@/hooks/my-page'
