+++ src/components/common/GaugeBar.tsx
@@ -1,6 +1,7 @@
-import JustifiedText from '@/components/common/gauge-bar/JustifiedText'
 import { tasteMetaMap } from '@/constants'
 import type { TasteScore } from '@/types/gauge-bar/tasteTypes'
+
+import JustifiedText from './gauge-bar/JustifiedText'
 
 const GaugeBar = ({ type, score }: TasteScore) => {
   const { label, color } = tasteMetaMap[type]
