+++ src/components/common/index.ts
@@ -24,3 +24,4 @@ export { default as EditNickNameModal } from './modal/EditNickNameModal'
 export { default as TasteReviewModal } from './modal/TasteReviewModal'
 export { default as SelectCartModal } from './modal/SelectCartModal'
 export { default as ReviewModal } from './modal/ReviewModal'
+export { default as ResignationModal } from './modal/ResignationModal'
